requirements.txt
asgiref==3.5.2
bson==0.5.10
dj-database-url==0.5.0
Django==4.0.4
django-heroku==0.3.1
djangorestframework==3.13.1
dnspython==2.2.1
gunicorn==20.1.0
psycopg2==2.9.3
pymongo==4.1.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.1.0
